+++ pixeltable/utils/arrow.py
@@ -1,6 +1,5 @@
 import datetime
 import io
-import json
 import uuid
 from typing import TYPE_CHECKING, Any, Iterator, cast
 
@@ -108,8 +107,7 @@ def to_arrow_schema(pxt_schema: dict[str, ts.ColumnType]) -> pa.Schema:
         arrow_type = to_arrow_type(col_type)
         if arrow_type is None:
             raise excs.RequestError(
-                excs.ErrorCode.UNSUPPORTED_OPERATION,
-                f'Cannot convert column {col_name!r} of type {col_type} to arrow.',
+                excs.ErrorCode.UNSUPPORTED_OPERATION, f'Cannot convert column {col_name!r} of type {col_type} to arrow.'
             )
         pa_column_types[col_name] = arrow_type
     return pa.schema(pa_column_types.items())
@@ -144,9 +142,6 @@ def _to_record_batch(
             # convert ragged arrays to nested lists
             list_col_vals = [val.tolist() for val in column_vals[field.name]]
             pa_arrays.append(pa.array(list_col_vals))
-        elif field.type == pa.json_():
-            serialized = [json.dumps(v) if v is not None else None for v in column_vals[field.name]]
-            pa_arrays.append(pa.array(serialized, type=pa.json_()))
         elif pxt_type.is_json_type():
             # JSON columns are typed by `pa.infer_type` against the first batch's values; that can
             # succeed (e.g. `list<int64>` inferred from `[1, 'a', 2]`) but still fail when pyarrow
@@ -276,6 +271,30 @@ def to_record_batches(query: 'pxt.Query', batch_size_bytes: int) -> Iterator[pa.
         yield record_batch
 
 
+def find_null_fields(schema: pa.Schema) -> list[str]:
+    """Return dotted paths of every `pa.null()`-typed field nested inside `schema`."""
+    paths: list[str] = []
+
+    def walk(arrow_type: pa.DataType, path: str) -> None:
+        if pa.types.is_null(arrow_type):
+            paths.append(path)
+        elif pa.types.is_struct(arrow_type):
+            for f in arrow_type:
+                walk(f.type, f'{path}.{f.name}')
+        elif (
+            pa.types.is_list(arrow_type)
+            or pa.types.is_large_list(arrow_type)
+            or pa.types.is_fixed_size_list(arrow_type)
+        ):
+            walk(arrow_type.value_type, f'{path}[]')
+        elif pa.types.is_map(arrow_type):
+            walk(arrow_type.item_type, f'{path}{{}}')
+
+    for field in schema:
+        walk(field.type, field.name)
+    return paths
+
+
 def to_pydict(batch: pa.Table | pa.RecordBatch) -> dict[str, list | np.ndarray]:
     """Convert a RecordBatch to a dictionary of lists
 
@@ -289,8 +308,6 @@ def to_pydict(batch: pa.Table | pa.RecordBatch) -> dict[str, list | np.ndarray]:
             if isinstance(col, pa.ChunkedArray):
                 col = col.combine_chunks()
             out[name] = list(cast(pa.FixedShapeTensorArray, col).to_numpy_ndarray())
-        elif col.type == pa.json_():
-            out[name] = [json.loads(v) if v is not None else None for v in col.to_pylist()]
         else:
             # for the rest, use pydict to preserve python types
             out[name] = col.to_pylist()
